requirements.txt
streamlit
pdf2image
docx2txt
pypdf2
reportlab
python-docx
openpyxl